subshader

+++ lib/src/Resource/ResourcesManager.h
@@ -13,7 +13,7 @@ namespace fm {
 class ResourcesManager : protected fm_system::NonCopyable 
 {
 	typedef std::map<std::string, std::shared_ptr<Resource>> MapOfResources;
-	typedef std::array<MapOfResources, RESOURCE_TYPE::LAST_RESOURCE> ArrayOfResources;
+	typedef std::array<MapOfResources, (size_t)RESOURCE_TYPE::LAST_RESOURCE> ArrayOfResources;
 public:
 
 
@@ -22,8 +22,8 @@ public:
 
     template <typename T> std::shared_ptr<T> getResource(const std::string& name)
 	{
-		MapOfResources::const_iterator it = resources[T::getType()].find(name);
-        if(it == resources[T::getType()].cend()) 
+		MapOfResources::const_iterator it = resources[(size_t)T::getType()].find(name);
+        if(it == resources[(size_t)T::getType()].cend())
 			return nullptr;
 
         return std::dynamic_pointer_cast<T>(it->second);
@@ -31,8 +31,8 @@ public:
 
 	template <typename T> std::shared_ptr<T> getResource(const fm::FilePath& inPath)
 	{
-		MapOfResources::const_iterator it = resources[T::getType()].find(fm::FileSystem::ConvertPathToFileSystem(inPath));
-		if (it == resources[T::getType()].cend())
+		MapOfResources::const_iterator it = resources[(size_t)T::getType()].find(fm::FileSystem::ConvertPathToFileSystem(inPath));
+		if (it == resources[(size_t)T::getType()].cend())
 			return nullptr;
 
 		return std::dynamic_pointer_cast<T>(it->second);
@@ -40,39 +40,39 @@ public:
 
     template <typename T> bool Exists(const std::string& name)
 	{
-        return resources[T::getType()].find(name) != resources[T::getType()].end();
+        return resources[(size_t)T::getType()].find(name) != resources[T::getType()].end();
     }
 
     template <typename T>
     void load(const std::string& name, std::shared_ptr<Resource> resource) 
 	{
         if(T::getType() > RESOURCE_TYPE::LAST_RESOURCE) return;
-        resources[T::getType()].insert(std::pair<std::string, std::shared_ptr<Resource>>(name, resource));
+        resources[(size_t)T::getType()].insert(std::pair<std::string, std::shared_ptr<Resource>>(name, resource));
     }
 
 	void load(const std::string& name, std::shared_ptr<Resource> resource)
 	{
 		if (resource->GetType() > RESOURCE_TYPE::LAST_RESOURCE) return;
-		resources[resource->GetType()].insert(std::pair<std::string, std::shared_ptr<Resource>>(name, resource));
+		resources[(size_t)resource->GetType()].insert(std::pair<std::string, std::shared_ptr<Resource>>(name, resource));
 	}
 
 	template <typename T>
 	void load(const fm::FilePath& inPath, std::shared_ptr<Resource> resource)
 	{
 		if (T::getType() > RESOURCE_TYPE::LAST_RESOURCE) return;
-		resources[T::getType()].insert(std::pair<std::string, std::shared_ptr<Resource>>(fm::FileSystem::ConvertPathToFileSystem(inPath), resource));
+		resources[(size_t)T::getType()].insert(std::pair<std::string, std::shared_ptr<Resource>>(fm::FileSystem::ConvertPathToFileSystem(inPath), resource));
 	}
 
 	void load(const fm::FilePath& inPath, std::shared_ptr<Resource> resource)
 	{
 		if (resource->GetType() > RESOURCE_TYPE::LAST_RESOURCE) return;
-		resources[resource->GetType()].insert(std::pair<std::string, std::shared_ptr<Resource>>(fm::FileSystem::ConvertPathToFileSystem(inPath), resource));
+		resources[(size_t)resource->GetType()].insert(std::pair<std::string, std::shared_ptr<Resource>>(fm::FileSystem::ConvertPathToFileSystem(inPath), resource));
 	}
     
     template <typename T>
     std::map<std::string, std::shared_ptr<Resource> >& getAll()
 	{
-        return resources[T::getType()];
+        return resources[(size_t)T::getType()];
     }
 
 	static FilePath GetFilePathResource(LOCATION inLocation);
